+++ src/main/java/com/accenture/flowershop/fe/dto/OrderDTO.java
@@ -3,13 +3,14 @@ package com.accenture.flowershop.fe.dto;
 import com.accenture.flowershop.be.entity.order.Item;
 import com.accenture.flowershop.fe.enums.order.Status;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
 public class OrderDTO {
     private int id;
     private int orderNum;
-    private double orderSubtotal;
+    private BigDecimal orderSubtotal;
     private List<Item> items;
     private Date orderDate;
     private Date orderCloseDate;
